[#176] This commit moves the Upvoter component out of Caption.js and into a separate file.
